Make the info button always visible on charity screen
